fix: edit all entity & repository

+++ backend/treeway/src/main/java/com/b107/treeway/api/article/repository/ArticleAttachedFileRepository.java
@@ -4,6 +4,6 @@ package com.b107.treeway.api.article.repository;
 import com.b107.treeway.api.article.entity.ArticleAttachedFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleAttachedFileRepository extends JpaRepository<ArticleAttachedFile, Integer> {
+public interface ArticleAttachedFileRepository extends JpaRepository<ArticleAttachedFile, Long> {
 
 }
