+++ backend/treeway/src/main/java/com/b107/treeway/api/article/repository/ArticleScrapRepository.java
@@ -4,6 +4,6 @@ package com.b107.treeway.api.article.repository;
 import com.b107.treeway.api.article.entity.ArticleScrap;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleScrapRepository extends JpaRepository<ArticleScrap, Integer> {
+public interface ArticleScrapRepository extends JpaRepository<ArticleScrap, Long> {
 
 }
